Implementation to read OSM files directly.
Implementation based on https://github.com/dkondor/osmconvert/, with simplifications.

It performs the following steps:

1. Read all OSM nodes in the input and read ways, but filter them to only include those that are suitable for car travel
(similar to the 'drive' filter in osmnx -- https://github.com/gboeing/osmnx/blob/46ed807f870e3bede78f7e494e98490df470bd42/osmnx/downloader.py#L19 )

2. Process ways and store a graph between the OSM nodes.

3. Process the graph by excluding OSM nodes with degree == 2 and creating edges
between nodes with degree != 2 along chains. Create a LineString for each edge
that contains the original geometry.

4. Add these edges to the FMM::NETWORK::Network class.

Limitations:

 - Edges are assigned arbitrary IDs during the processing. This is necessary
since OSM way IDs are not guaranteed to be unique (a way object can span
multiple intersections). This way, the IDs in the output are meaningless.
There should be some way to avoid this problem, either by outputting the
assigned edge IDs separately or maybe outputting the OSM node IDs in the path
instead of edge IDs, which are indeed unique and can be used to identify the
path later.

 - Filtering for ways is not consistent with osmnx -- way_filter class could be
modified to reproduce the behavior of e.g. the 'drive' filter. Also, there could
be an option to avoid filtering if it was done by an external program before.

 - It is not possible to cut the network; this needs to be done externally as
well. Also, no attempt is made to detect connected components; this might be
better done inside the Network class.

+++ src/network/type.hpp
@@ -19,7 +19,7 @@
 namespace FMM {
 namespace NETWORK{
 
-typedef int NodeID; /**< Node ID in the network, can be discontinuous int */
+typedef uint64_t NodeID; /**< Node ID in the network, can be discontinuous int */
 typedef int EdgeID; /**< Edge ID in the network, can be discontinuous int */
 typedef unsigned int NodeIndex; /**< Node Index in the network, range
                                  from [0,num_vertices-1 ]*/
